Move all files from lib to src
