+++ src/controllers/logout.js
