+++ src/index.test-d.ts
